test: Verify that MinGW binaries may be run via Wine
Patch the Wine system.reg file to include the path to the MinGW
libraries in order to allow cross-compiled binaries to be executed,
and verify that executing a binary succeeds.

Note that binfmt_misc cannot (sanely) be used within the Fedora
container, and so the binaries may not simply be invoked as normal
executables.  This will prevent the use of "make check".s

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>