common: Separate aligned and unaligned memory allocation
The posix_memalign() function guarantees that allocated memory can be
freed using the standard free() library call.  This guarantee is not
provided for _aligned_malloc(), which requires the use of
_aligned_free() in place of free().

We need any memory that UEFI states should be freed using FreePool()
to be freeable using the standard free() library call, since this
avoids the need for consuming code to be aware of any custom memory
allocation mechanism.  This in turn implies that our implementation of
AllocatePool() must produce pointers that can be passed to free().
Since _aligned_malloc() does not have this property, we must not use
_aligned_malloc() within the implementation for AllocatePool().

Fix by separating out concepts of aligned and unaligned memory
allocation in our MemoryAllocationLib implementation.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>