build: Add support for generating code coverage reports
Enable the generation of code coverage reports via AX_CODE_COVERAGE.
A coverage report may be generated using

  ./configure --enable-code-coverage
  make check-code-coverage

The usage of AX_CODE_COVERAGE was changed in autoconf-archive
v2019.01.06 in a backwards-incompatible way.  The contortions required
to support both versions are ugly (and against the philosophy of using
autotools).

One possible solution would be to include a current version of
ax_code_coverage.m4 and all of its dependencies within this
repository.  This would add a substantial amount of noise to the
codebase, and would potentially introduce compatibility issues with
other macros (such as AX_CHECK_LINK_FLAG) obtained from the local
system's autoconf-archive installation.

Since autoconf-archive v2019.01.06 or later is included in current
releases of both Fedora and Ubuntu, a cleaner solution seems to be to
require the use of the ubuntu-20.04 runner in the initial GitHub
workflow.

This issue affects only builds direct from the git repository.  Builds
from the distribution tarball (including subsequent jobs in the GitHub
workflow) do not rerun autogen.sh and so do not need any version of
autoconf-archive available.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>