efidevpath: Add concept of an implausible device path
When constructing a device path from a textual representation, EFI
will treat any unrecognised string as a filename.  This can lead to
unexpected behaviour.  For example, if a device path component Uri()
is mistyped as URI(), it will become a FILEPATH_DEVICE_PATH with the
filename "URI()", rather than becoming a URI_DEVICE_PATH.

It is implausible that any real FILEPATH_DEVICE_PATH would be of the
form "Xxx(...)".  Add a function to check for such implausibility.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>